feat: 리뷰 수정 기능 구현

+++ backend/src/main/java/com/carffeine/carffeine/station/controller/review/dto/ReviewResponse.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 public record ReviewResponse(Long reviewId,
                              Long memberId,
                              LocalDateTime latestUpdateDate,
-                             int ratings,
+                             long ratings,
                              String content,
                              boolean isUpdated,
                              boolean isDeleted) {
